Change RollResult to use doubles

+++ src/Frags.Presentation/Results/RollResult.cs
@@ -19,7 +19,7 @@ namespace Frags.Presentation.Results
         /// <param name="name">The character name.</param>
         /// <param name="stat">The rolled statistic.</param>
         /// <param name="rollValue">What the character rolled.</param>
-        public static RollResult Roll(string name, string stat, int rollValue) =>
+        public static RollResult Roll(string name, string stat, double rollValue) =>
             new RollResult($"{name} rolled a {rollValue} in {stat}.");
 
         /// <summary>
@@ -30,7 +30,7 @@ namespace Frags.Presentation.Results
         /// <param name="roll1">The first character's roll.</param>
         /// <param name="roll2">The second character's roll.</param>
         public static RollResult RollAgainst(
-            string character1, string character2, int roll1, int roll2)
+            string character1, string character2, double roll1, double roll2)
         {
             if (roll1 > roll2)
                 return new RollResult(
